+++ requests_cache/models/__init__.py
@@ -9,7 +9,8 @@ from .base import RichMixin
 from .raw_response import CachedHTTPResponse
 from .request import CachedRequest
 from .response import CachedResponse, DecodedContent, OriginalResponse
+from .streaming import DeferredCacheResponse
 
-AnyResponse = Union[OriginalResponse, CachedResponse]
+AnyResponse = Union[OriginalResponse, CachedResponse, DeferredCacheResponse]
 AnyRequest = Union[Request, PreparedRequest, CachedRequest]
 AnyPreparedRequest = Union[PreparedRequest, CachedRequest]